Move AWSServiceErrors declaration to a better place
