+++ src/toil/provisioners/aws/awsProvisioner.py
@@ -45,8 +45,7 @@ from mypy_boto3_ec2.service_resource import Instance
 from mypy_boto3_iam.type_defs import InstanceProfileTypeDef, RoleTypeDef, ListRolePoliciesResponseTypeDef
 from mypy_extensions import VarArg, KwArg
 
-from toil.jobStores.aws.utils import AWSServerErrors
-from toil.lib.aws import zone_to_region, AWSRegionName
+from toil.lib.aws import zone_to_region, AWSRegionName, AWSServerErrors
 from toil.lib.aws.ami import get_flatcar_ami
 from toil.lib.aws.iam import (CLUSTER_LAUNCHING_PERMISSIONS,
                               get_policy_permissions,
